refactor(tests): Move server setup to base test case subclass
